vector-processor/requirements.txt
# Core dependencies for unified content processor
openai>=1.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
python-frontmatter>=1.0.0

# Legacy ChromaDB dependencies (for reference)
chromadb>=0.4.0
tiktoken>=0.5.0
gitpython>=3.1.0
pathspec>=0.11.0
requests>=2.28.0
pymongo>=4.0.0